requirements.txt
kucoin-python
pandas
numpy
scipy
psycopg2-binary
python-dotenv 